cambio de tinyint a int

+++ src/main/resources/db/migration/V20190327160428__Agregar_genero_a_Alumno.sql
@@ -1,1 +1,1 @@
-alter table Alumno modify genero tinyint(2);
+alter table Alumno modify genero int(11);
